file_export: moved MRLE video codec to its own file
