+++ src/Makefile.am
@@ -128,7 +128,7 @@ if WITH_MULTIMEDIA
 atari800_SOURCES += multimedia.c multimedia.h
 endif
 if WITH_FILE_EXPORT
-atari800_SOURCES += file_export.c file_export.h
+atari800_SOURCES += file_export.c file_export.h video_codec_mrle.c video_codec_mrle.h
 endif
